Move aliasUtils to alias namespace
